Move all metrics to the same file
